refactor: move into separate submodules for cleaner organization
